requirement.txt
Flask
gunicorn
langdetect
google-generativeai
reportlab